requirements.txt
torch==2.5.1
torchvision==0.20.1
segmentation-models-pytorch==0.2.1
numpy==1.26.4
matplotlib==3.9.2
tqdm==4.66.5
Pillow==10.4.0
onnx==1.17.0
onnxruntime==1.20.0
